move learn svgs into tools/learn/components/svgs
