+++ src-ts/lib/svgs/index.ts
@@ -3,9 +3,6 @@ import * as IconOutline from '@heroicons/react/outline'
 import * as IconSolid from '@heroicons/react/solid'
 
 import { ReactComponent as ActiveTabTipIcon } from './activetab-tip-icon.svg'
-import { ReactComponent as LearnChallengeBadgeIcon } from './learn-challenge-badge-icon.svg'
-import { ReactComponent as LearnGetCertified } from './learn-get-certified.svg'
-import { ReactComponent as LearningHat } from './learning-hat.svg'
 import { ReactComponent as LogoIcon } from './logo.svg'
 import { ReactComponent as SocialIconFacebook } from './social-fb-icon.svg'
 import { ReactComponent as SocialIconInstagram } from './social-insta-icon.svg'
@@ -20,9 +17,6 @@ import { ReactComponent as WorkTypeCategoryUnknownIcon } from './work-type-categ
 export { ActiveTabTipIcon }
 export { IconOutline }
 export { IconSolid }
-export { LearnChallengeBadgeIcon }
-export { LearnGetCertified }
-export { LearningHat }
 export { LogoIcon }
 export { SocialIconFacebook }
 export { SocialIconInstagram }
