+++ src-ts/tools/learn/components/index.ts
@@ -2,4 +2,5 @@ export * from './collapsible-pane'
 export * from './course-outline'
 export * from './course-title'
 export * from './status-checkbox'
+export * from './svgs'
 export * from './tc-academy-policy-modal'
