+++ src-ts/tools/learn/components/svgs/index.ts
@@ -1,0 +1,7 @@
+import { ReactComponent as LearnChallengeBadgeIcon } from './learn-challenge-badge-icon.svg'
+import { ReactComponent as LearnGetCertified } from './learn-get-certified.svg'
+import { ReactComponent as LearningHat } from './learning-hat.svg'
+
+export { LearnChallengeBadgeIcon }
+export { LearnGetCertified }
+export { LearningHat }
