+++ src-ts/tools/learn/components/svgs/learn-welcome-bg-curve.png
