+++ src-ts/tools/learn/welcome/WelcomePage.module.scss
@@ -3,7 +3,7 @@
 
 .hero {
     &-wrap {
-        background: url('../../../lib/svgs/learn-welcome-bg-curve.png') repeat-x center bottom , $tc-grad15;
+        background: url('../components/svgs/learn-welcome-bg-curve.png') repeat-x center bottom , $tc-grad15;
         background-size: 1440px 60px, auto;
         overflow: hidden;
         padding-bottom: 60px;
