+++ src-ts/tools/learn/welcome/progress-block/ProgressBlock.tsx
@@ -1,8 +1,9 @@
 import { FC, ReactNode, useCallback } from 'react'
 import { NavigateFunction, useNavigate } from 'react-router-dom'
 
-import { Button, LearningHat } from '../../../../lib'
+import { Button } from '../../../../lib'
 import { LearnMyCertificationProgress, MyCertificationsProviderData, useMyCertifications } from '../../services'
+import { LearningHat } from '../../components'
 
 import { Completed } from './completed'
 import { InProgress } from './in-progress'
